Making progress towards getting TypeScript Working
