+++ src/styles/style.scss
@@ -181,10 +181,9 @@ button {
 }
 
 .chartCanvas span {
-  display: flex;
   height: 100%;
-  align-items: center;
-  justify-content: center;
+  display: grid;
+  place-items: center;
 }
 
 #metrics footer {
