+++ tsconfig.json
@@ -17,10 +17,14 @@
 // ]
 // }
 
+// Added from the redux template https://github.com/reduxjs/redux-templates/blob/master/packages/cra-template-redux-typescript/tsconfig.json
+// module and moduleResolution may require extra reading at:
+// https://stackoverflow.com/questions/71463698/why-we-need-nodenext-typescript-compiler-option-when-we-have-esnext#:~:text=The%20biggest%2C%20most%20noticeable%20difference,because%20there%20is%20no%20such
+
 {
   "compilerOptions": {
     "target": "ES2020",
-    "module": "NodeNext",
+    "module": "ESNext",
     "jsx": "react",
     "allowJs": true,
     "outDir": "./build",
@@ -33,17 +37,17 @@
     "forceConsistentCasingInFileNames": true,
     "noEmit": true,
     "allowImportingTsExtensions": true,
-    "types": ["webpack-env"]
+    "types": ["webpack-env"],
+    "lib": ["dom", "dom.iterable", "esnext"],
+    "allowSyntheticDefaultImports": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true
   },
   "include": [
     "src/**/*.ts",
     "src/**/*.tsx",
     "ts-server/**/*.ts",
-    "ts-server/**/*.tsx",
-    "src/redux/store.js",
-    "src/containers/Clusters.jsx",
-    "src/containers/Main.jsx",
-    "src/reducers/clusterSlice.js"
+    "ts-server/**/*.tsx"
   ],
   "exclude": ["node_modules"]
 }
