+++ webpack.config.js
@@ -8,7 +8,7 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 export default {
-  entry: './src/index.jsx',
+  entry: './src/index.tsx',
 
   output: {
     path: path.join(__dirname, '/dist'),
